add a Container class for the DI
