+++ Slim/Environment.php
@@ -48,12 +48,8 @@ namespace Slim;
  * @author  Josh Lockhart
  * @since   1.6.0
  */
-class Environment implements \ArrayAccess, \IteratorAggregate
+class Environment extends Container implements \IteratorAggregate
 {
-    /**
-     * @var array
-     */
-    protected $properties;
 
     /**
      * @var \Slim\Environment
@@ -187,42 +183,6 @@ class Environment implements \ArrayAccess, \IteratorAggregate
 
             $this->properties = $env;
         }
-    }
-
-    /**
-     * Array Access: Offset Exists
-     */
-    public function offsetExists($offset)
-    {
-        return isset($this->properties[$offset]);
-    }
-
-    /**
-     * Array Access: Offset Get
-     */
-    public function offsetGet($offset)
-    {
-        if (isset($this->properties[$offset])) {
-            return $this->properties[$offset];
-        } else {
-            return null;
-        }
-    }
-
-    /**
-     * Array Access: Offset Set
-     */
-    public function offsetSet($offset, $value)
-    {
-        $this->properties[$offset] = $value;
-    }
-
-    /**
-     * Array Access: Offset Unset
-     */
-    public function offsetUnset($offset)
-    {
-        unset($this->properties[$offset]);
     }
 
     /**
