+++ tests/bootstrap.php
@@ -1,10 +1,8 @@
 <?php
 set_include_path(dirname(__FILE__) . '/../' . PATH_SEPARATOR . get_include_path());
 
-require_once 'Slim/Slim.php';
-
-// Register Slim's autoloader
-\Slim\Slim::registerAutoloader();
+require 'Slim/Autoloader.php';
+\Slim\Autoloader::registerAutoloader();
 
 //Register non-Slim autoloader
 function customAutoLoader( $class )
